Rename global to hyperion-global package
Rename the package to something more unique to allow us release to npm
later.
