+++ packages/hyperion-autologging/package.json
@@ -13,13 +13,13 @@
   },
   "workspaces": [
     "../hyperion-devtools",
-    "../global",
+    "../hyperion-global",
     "../hook",
     "../hyperion-dom",
     "../hyperion-react"
   ],
   "dependencies": {
-    "@hyperion/global": "^0.1.0",
+    "@hyperion/hyperion-global": "^0.1.0",
     "@hyperion/hook": "^0.1.0",
     "@hyperion/hyperion-dom": "^0.1.0",
     "@hyperion/hyperion-react": "^0.1.0"
