+++ packages/hyperion-autologging/src/ALTriggerFlowlet.ts
@@ -6,7 +6,7 @@
 import type * as Types from "@hyperion/hyperion-util/src/Types";
 import type * as React from 'react';
 
-import { assert } from "@hyperion/global";
+import { assert } from "@hyperion/hyperion-global";
 import { Channel } from "@hyperion/hook/src/Channel";
 import { getFunctionInterceptor, interceptFunction } from "@hyperion/hyperion-core/src/FunctionInterceptor";
 import { getVirtualPropertyValue, setVirtualPropertyValue } from "@hyperion/hyperion-core/src/intercept";
