+++ packages/hyperion-core/package.json
@@ -12,11 +12,11 @@
   },
   "workspaces": [
     "../hyperion-devtools",
-    "../global",
+    "../hyperion-global",
     "../hook"
   ],
   "dependencies": {
-    "@hyperion/global": "^0.1.0",
+    "@hyperion/hyperion-global": "^0.1.0",
     "@hyperion/hook": "^0.1.0"
   },
   "devDependencies": {
