+++ packages/hyperion-dom/src/DOMShadowPrototype.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
 
-import { assert } from "@hyperion/global";
+import { assert } from "@hyperion/hyperion-global";
 import { ShadowPrototype } from "@hyperion/hyperion-core/src/ShadowPrototype";
 import { VirtualAttribute } from "./VirtualAttribute";
 import { getObjectExtension } from "@hyperion/hyperion-core/src/intercept";
