+++ packages/hyperion-dom/src/IElementCustom.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
  */
 
-import { assert } from "@hyperion/global";
+import { assert } from "@hyperion/hyperion-global";
 import { getVirtualAttribute } from "./DOMShadowPrototype";
 import * as IElement from "./IElement_";
