+++ packages/hyperion-global/src/AsyncCounter.ts
