+++ packages/hyperion-global/src/assert.ts
