+++ packages/hyperion-global/src/global.d.ts
