+++ packages/hyperion-global/src/global.ts
