+++ packages/hyperion-react/src/IReactComponent.ts
@@ -13,7 +13,7 @@ import type {
   ReactSpecialComponentTypes
 } from './IReact';
 
-import { assert } from '@hyperion/global';
+import { assert } from '@hyperion/hyperion-global';
 import { Hook } from '@hyperion/hook/src/Hook';
 import { FunctionInterceptor, interceptFunction } from '@hyperion/hyperion-core/src/FunctionInterceptor';
 import { interceptMethod } from '@hyperion/hyperion-core/src/MethodInterceptor';
